+++ src/routes/github/webhook/webhook-receiver-post.ts
@@ -2,7 +2,7 @@ import Logger from "bunyan";
 import { BinaryLike, createHmac } from "crypto";
 import { Request, Response } from "express";
 import { getLogger } from "~/src/config/logger";
-import { GitHubServerApp } from "~/src/models/git-hub-server-app";
+import { GitHubServerApp } from "models/github-server-app";
 import { WebhookContext } from "./webhook-context";
 
 export const WebhookReceiverPost = async (request: Request, response: Response): Promise<void> => {
