+++ src/util/get-github-client-config.ts
@@ -1,6 +1,6 @@
 // this will need to be updated later to check for database entry
 import { Installation } from "models/installation";
-import { GitHubServerApp } from "models/git-hub-server-app";
+import { GitHubServerApp } from "models/github-server-app";
 import { GitHubInstallationClient } from "../github/client/github-installation-client";
 import { getCloudInstallationId } from "../github/client/installation-id";
 import { booleanFlag, BooleanFlags } from "config/feature-flags";
